Updates the authentication process to support passwordCredentials
This change adds some structure to the auth process to ensure that
we're able provide access to a cloud using both the password or
the apikey from Rackspace Identity. The auth method will check,
if the provided password value is a valid hash and if so, it use
the API key authentication method first. In the event of a failure
the method will retry using the password authentication method.
If the hash validation check fails the plugin will assume the
user is attempting basic password authentication.

This change also adds input validation and logging so that we're
better able to debug the plugin in the event of an issue.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>