ignore non-OpenStack Rackspace products (#2)
The Rackspace service catalog returns roles for non-OpenStack products
that Rackspace offers. These do not using the same types of values or
mappings that OpenStack uses which causes an issue. They aren't
something that would be supported in keystone anyway since they exist
outside of identity for roles and permissions so ignore them from this
plugin.

Signed-off-by: Doug Goldstein <doug.goldstein@rackspace.com>