feat: update keystone-rxt to use the new rxt role
This change makes it possible to run the keystone rxt plugin in
an environment using the os_flex attribute.

This change includes
* multi-account detection and autoprovisioning
* new mapping defaults for scheme version 2.0

Two new options have been added to support the creation of the
role attribute methods, allowing a user to be a member of more
than one project.

[rackspace]
role_attribute_enforcement <BOOL>
role_attribute <STR>

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>