feat: ensure that projects have consistent ids
This change adjusts how shadow projects are stored in the database.
When a project is created, it is looked up using the name and when
not found keystone would normally assign it a random UUID. This
ensures that all projects have a consistent hash when being assigned
which will ensure that project IDs remain consistent between regions.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>