fix(project_id): fix hashing
The project length needs to retain a limit of 32 to keep compatibility
with openstack projects. While the SHA224 length is allowed, services
like Octavia have internal limitations that are incompatible with the
allowed keystone drivers. This change updates our hashing so that we're
using a supported algo but limiting how we hash to a maximum length of
32.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>